MockWorld configurable VMType
